4.21

+++ fullstack_open_exercises/part4/controllers/blogs.js
@@ -45,7 +45,17 @@ blogsRouter.get("/:id", async (request, response, next) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: "token invalid" });
+	}
+
 	try {
+		const blog = await Blog.findById(request.params.id);
+		if (blog.user.toString() !== decodedToken.id.toString()) {
+			return response.status(403).json({ error: "no permissions" });
+		}
+
 		await Blog.findByIdAndDelete(request.params.id);
 		response.status(204).end();
 	} catch (error) {
